+++ Omega/src/com/saggitt/omega/settings/search/SettingsSearchView.kt
@@ -31,7 +31,7 @@ class SettingsSearchView(context: Context, attrs: AttributeSet?) : SearchView(co
     private var customTypeface: Typeface? = null
 
     init {
-        isIconified = false;
+        isIconified = false
     }
 
     fun setTypeface(typeface: Typeface) {