+++ src_flags/com/android/launcher3/config/FeatureFlags.java
@@ -23,9 +23,7 @@ public final class FeatureFlags extends BaseFlags {
 
     public static boolean REFLECTION_FORCE_OVERVIEW_MODE = true;
 
-    // When enabled add space for a search widget in the dock.
-    public static boolean HOTSEAT_WIDGET = true;
-
+    public static boolean FORCE_FEED_BRIDGE = false;
     private FeatureFlags() {
         // Prevent instantiation
     }